added EM and 2 NVT stages

+++ src/pmx/scripts/workflows/Abs_dG_lig_binding.py
@@ -99,7 +99,8 @@ def parse_options():
 class Workflow_inProtein:
     def __init__(self, toppath, mdppath, proteins=[], ligands=[],
                  n_repeats=3, n_sampling_sims=1, basepath=os.getcwd(),
-                 d=1.5, bt="dodecahedron", salt_conc=0.15):
+                 d=1.5, bt="dodecahedron", salt_conc=0.15,
+                 mdrun="gmx mdrun", mdrun_opts=""):
         self.toppath = toppath
         self.mdppath = mdppath
         self.n_repeats = n_repeats
@@ -110,6 +111,10 @@ class Workflow_inProtein:
         self.d = d
         self.bt = bt
         self.salt_conc = salt_conc
+        self.mdrun = mdrun
+        self.mdrun_opts = mdrun_opts
+        self.states={"A":"l0", "B":"l1"} #states and suffixes of mdp files
+        
         
     def check_sanity(self):
         if(not sh.which('gmx')):
@@ -163,7 +168,7 @@ class Workflow_inProtein:
                           "-fc 9000 9000 9000 "
                           "-o lig_posre.itp >> setup.log  2>&1")
                 check_file_ready("lig_posre.itp")
-                os.system("echo 'MOL\n' | gmx genrestr -f init.pdb "
+                os.system("echo 'MOL\n' | gmx genrestr -f lig.pdb "
                           "-fc 500 500 500 "
                           "-o lig_posre_soft.itp >> setup.log 2>&1")
                 check_file_ready("lig_posre_soft.itp")
@@ -175,7 +180,7 @@ class Workflow_inProtein:
                         try:
                             os.unlink(filePath)
                         except:
-                            print("Error while deleting file : ", filePath)
+                            print("Error while deleting file: ", filePath)
                 
                 #Return to basepath
                 os.chdir(self.basepath)
@@ -186,28 +191,32 @@ class Workflow_inProtein:
                 folder = self.gen_folder_name(p,l)
                 os.chdir(folder)
                 
-                #generate the box
-                os.system("gmx editconf -f init.pdb -o box.pdb -bt %s -d %f "\
-                          "> prep.log 2>&1"%(self.bt, self.d))
-                check_file_ready("box.pdb")
-                os.system("gmx solvate -scale 1.0 -cp box.pdb -o water.pdb "\
-                          "-cs spc216.gro -p topol.top >> prep.log 2>&1")
-                check_file_ready("water.pdb")
-                os.system("gmx grompp -p topol.top -c water.pdb -o tpr.tpr "\
-                          "-f %s/protein/init.mdp -v -maxwarn 2 "\
-                          ">> prep.log 2>&1"%self.mdppath)
-                check_file_ready("tpr.tpr")
+                #generate the box and solvate
+                if(not os.path.isfile("tpr.tpr")): #skip if it already exists
+                    os.system("gmx editconf -f init.pdb -o box.pdb -bt %s -d %f "\
+                              "> prep.log 2>&1"%(self.bt, self.d))
+                    check_file_ready("box.pdb")
+                    os.system("gmx solvate -scale 1.0 -cp box.pdb -o water.pdb "\
+                              "-cs spc216.gro -p topol.top >> prep.log 2>&1")
+                    check_file_ready("water.pdb")
+                    os.system("gmx grompp -p topol.top -c water.pdb -o tpr.tpr "\
+                              "-f %s/protein/init.mdp -v -maxwarn 2 "\
+                              ">> prep.log 2>&1"%self.mdppath)
+                    check_file_ready("tpr.tpr")
                 
-                #independent repeats for error analysis
+                #generate ions for each
+                #independent repeat (multiple for confidence estimate)
                 for i in range(self.n_repeats):
                     #sampling simulations in each repeat
                     for m in range(self.n_sampling_sims):
                         top_ions="topol_ions%d_%d.top"%(i,m)
                         pdb_ions="ions%d_%d.pdb"%(i,m)
+                        if(os.path.isfile(pdb_ions)): #skip if it already exists
+                            continue
                         sh.copy("topol.top", top_ions)
                         os.system("echo 'SOL' | gmx genion -s tpr.tpr "
                                   "-p %s -conc %f "
-                                  "-neutral -nname ClJ -pname NaJ "
+                                  "-neutral -nname Cl -pname Na "
                                   "-o %s >> genion.log 2>&1" %(
                                       top_ions, self.salt_conc, pdb_ions) )
                         check_file_ready(pdb_ions)
@@ -219,10 +228,72 @@ class Workflow_inProtein:
                         try:
                             os.unlink(filePath)
                         except:
-                            print("Error while deleting file : ", filePath)
+                            print("Error while deleting file: ", filePath)
                 
                 #Return to basepath
                 os.chdir(self.basepath)
+                
+    def run_stage(self, stage, mdp_template, struct_template,
+                  posre=None, clean=True):
+        print("Running stage "+stage+":")
+        for p in self.proteins:
+            for l in self.ligands:
+                folder = self.gen_folder_name(p,l)
+                
+                #independent repeats for error analysis
+                for i in range(self.n_repeats):
+                    #sampling simulations in each repeat
+                    for m in range(self.n_sampling_sims):
+                        for s in self.states:
+                            sim_folder=folder+"/state%s/repeat%d/%s%d"%(s,i,stage,m)
+                            os.makedirs(sim_folder, exist_ok=True)
+                            os.chdir(sim_folder)
+                            print("\t"+sim_folder)
+                            #don't rerun if already ready
+                            if(os.path.isfile("confout.gro")):
+                                print("\t\tPrevious")
+                                continue
+                            
+                            #"../data/mdp/em_{0}.mdp"
+                            mdp=mdp_template.format(self.states[s])#insert the l0/l1 suffix
+                            #"prot_{0}/lig_{1}/ions{3}_{4}.pdb"
+                            #"prot_{0}/lig_{1}/state{2}/repeat{3}/em{4}/confout.gro"
+                            struct=struct_template.format(p,l,s,i,m)#insert the s,i, and m suffixes
+                            top_template=folder+"/topol_ions{3}_{4}.top"
+                            top=top_template.format(p,l,s,i,m)
+                            
+                            #restraint
+                            r=""
+                            if posre:
+                                r="-r "+posre.format(p,l,s,i,m)
+                            
+                            #make tpr
+                            os.system("gmx grompp -p %s -c %s %s "
+                                      "-o tpr.tpr -f %s -v -maxwarn 3 "
+                                      "> prep.log 2>&1"%(
+                                          top, struct, r, mdp)
+                                      )
+                            check_file_ready("tpr.tpr")
+                            
+                            #run sim
+                            os.system(self.mdrun+" -s tpr.tpr "+\
+                                      self.mdrun_opts+" > mdrun.log 2>&1")
+                            check_file_ready("confout.gro")
+                            print("\t\tDone")
+                            
+                            
+                            #clean overwritten files
+                            if(clean):
+                                cleanList = glob.glob(sim_folder+'/#*')
+                                for filePath in cleanList:
+                                    try:
+                                        os.unlink(filePath)
+                                    except:
+                                        print("Error while deleting file: ", filePath)
+                            
+                            #Return to basepath
+                            os.chdir(self.basepath)
+                        
 
 # ==============================================================================
 #                               FUNCTIONS
@@ -238,9 +309,12 @@ def main(args):
     """
     toppath=os.path.abspath(args.toppath)
     mdppath=os.path.abspath(args.mdppath)
+    basepath=os.getcwd()
     
     w=Workflow_inProtein(toppath, mdppath, ["BRD1"], ["lig"],
-                         basepath=os.getcwd())
+                         basepath=basepath,
+                         mdrun="mdrun_threads_AVX2_256",
+                         mdrun_opts="-pin on -nsteps 10")
     
     #sanity checks
     w.check_sanity()
@@ -253,10 +327,19 @@ def main(args):
     w.prep()
     
     #run EM
-    
+    w.run_stage("em", mdppath+"/protein/em_posre_{0}.mdp",
+                basepath+"/prot_{0}/lig_{1}/ions{3}_{4}.pdb",
+                posre=basepath+"/prot_{0}/lig_{1}/ions{3}_{4}.pdb")
+        
     #run NVT w hard position restraints to prevent protein deformation
+    w.run_stage("nvt_posre", mdppath+"/protein/eq_nvt_posre_{0}.mdp",
+                basepath+"/prot_{0}/lig_{1}/state{2}/repeat{3}/em{4}/confout.gro",
+                posre=basepath+"/prot_{0}/lig_{1}/ions{3}_{4}.pdb")
     
     #run NVT w softer position restraints
+    w.run_stage("nvt_posre_soft", mdppath+"/protein/eq_nvt_posre_soft_{0}.mdp",
+                basepath+"/prot_{0}/lig_{1}/state{2}/repeat{3}/nvt_posre{4}/confout.gro",
+                posre=basepath+"/prot_{0}/lig_{1}/ions{3}_{4}.pdb")
     
     #run NPT to sample starting frames for TI
